[caldav] Decrease the window of possible missed updates
Previously, the plugin added thirty seconds to the previously recorded
last syccessful sync time, and used that timestamp when determining
which changes had occurred since last sync.  This meant that there
existed a window of 30 seconds within which updates would be "lost"
and not detected during the sync process.

This commit reduces the window down to 2 seconds (selected to be long
enough to allow database commit fs syncs to complete, but short enough
to reduce the chance of missed updates by a very significant factor).