Don't do a GET after a PUT as it will fail, and the client follows up with a REPORT anyway
The GET is only required to retrieve the updated etags of the files
sent by the PUT, and this is not necessary since caldavclient sends a
REPORT at the end of each sync, which will cause it to receive the
updated etags without an extra GET request.

Additionally, the PUT was always failing as its result depended on
the following GET, which fails as it attempts to read the received
body data in xml instead of vcal format. This patch changes the
GET response processing code to always emit finished().