Update last sync time when a sync finishes, and remove syncFinished() signal
Whenever a sync ends, mResults must be reset with the current
date/time so that this will be set by the buteo framework as the new
lastSyncTime in the sync log. Otherwise, the lastSyncTime will refer to
the time at which the sync session started, which means any events
received during the session will be seen as "new" events during the next
sync session, and will cause the client to try to send these as new
events to the server, which will fail.

This also renames receiveSyncFinished() to syncFinished() and removes
the existing syncFinished() signal as it was only used internally and
was always forwarded to receiveSyncFinished().