[caldav] Update recurrence and all-day data correctly
Recurrence data needs to be copied using Incidence serialization, as
the incidence's internal recurrence data will not be recreated
otherwise.

All-day information must have a time of midnight on one day to
midnight on the next, else the incidence is not recognized as being
an all-day event otherwise, even if the flag is set.