[caldav] Set exclusive DTEND when exporting events to the server
iCalendar format states DTEND must be exclusive. This must be set
else multi-day all-day events will not have the correct dates.
For single-day all-day events this wasn't a problem previously as
the dtEnd was not set at all, but it should be set anyway in this
case to be consistent.