[caldav] Save etags returned from PUT requests, and save them to the database instead of as custom incidence properties
Some servers update the event etag as soon as a PUT update request
is received for that event, and send back the updated etag immediately
in the PUT response. This etag must be saved, else it won't match the
locally saved etag and on the next sync the changed etag will make the
client incorrectly believe that the event has been modified.

However this means etags must now be saved to the database instead of
simply as a custom incidence property, as a custom property change is
still seen as an incidence modification by mkcal and thus it would
report this as a local modification on the next sync, causing event
updates to be sent unnecessarily.