Extract set of UIDs of all received incidences for later processing
In general, there is no correlation between the UID of an event and URI.
See, e.g. http://sabre.io/dav/building-a-caldav-client:

'The url and the UID have no meaningful relationship, so treat both
those items as separate unique identifiers.'

Thus, we should not use Reader::hrefToUid but use the real UIDs.