Correctly handle local deletions of incidences which were added during last sync
Incidences which have been received during the last sync and then were deleted locally
will not show up in localDeleted as their creation time is after the start time of
the last sync. Therefore, we have to go through the list of additions and add all
of them which are missing in the storage to the localDeleted list. Since
additions only holds UIDs and not incidences, make localDeleted a
QStringList.