Reload incidences sent to server in case we do not obtain an ETag
Some caldav servers, i.e. owncloud, do not return an ETag after a
successful PUT. In fact, this is not required by the specification (see
RFC 4791, section 5.3.4). This commit ensures that such incidences are
reloaded from the server using an additional multiget request.

Otherwise, any local modification to an incidence which has been sent to
the server during the last sync will be overwritten: we will see a
changed ETag on the server, and the server always wins in case of sync
conflicts ..