Use correct URI for DELETE requests.
We cannot use incidence.customProperty("buteo", "uri") since the storage
backend removes all custom properties from deleted incidences. We also
cannot use its UID as there is, in general, no correlation between the
UID of an incidence and its URI on the server.
However, we only DELETE an incidence if it is still on the server, and
all locally deleted incidences which are still on the server are
downloaded by fetchRemoteChanges. Therefore we can determine the URI by
searching mReceivedCalendarResources for the deleted incidence.