[caldav] Improve discovery of local deletions
Locally deleted incidences which have been received from the server
during the last sync are not reported by mStorage->deletedIncidences.
This is due to the fact that such incidences have a creation date after
the start date of the last sync. deletedIncidences only reports
incidences which have been deleted after a given date but have been
created before this date. As a result, the incidences in question are
retrieved as new incidences from the server instead of being deleted on
the server.

To fix this problem, we go through the list of remote additions during
the last sync and look for uids which have disappeared from the local
storage. These are then added to the list of local deletions.