[buteo-sync-plugins-caldav] Support detached occurrence (RECURRENCE-ID)
This commit adds support for syncing modifications to individual
occurrences of recurring event series.  These events have the same
UID as the original recurring event, but have a RECURRENCE-ID which
identifies that particular occurrence.

In order to support this, we require changes to the way we handle
every uid (by including uid+rid as the "true" identifier).