[buteo-sync-plugin-caldav] Prevent dtEnd data from being lost for all-day events
Don't copy the hasEndDate property in copyIncidenceProperties().
If the dtStart == dtEnd, calling setHasEndDate() causes the previous
setDtEnd() value to be lost and the dtEnd just becomes an invalid
KDateTime.

It's not necessary to copy hasEndDate anyway since internally this
will be set when setDtEnd() is called. For this reason, also remove
the call to setHasEndDate() in incidenceToExport(), since it already
calls setDtEnd().