[buteo-sync-plugin-caldav] Handle spaces in calendar paths. Contributes to MER#982
We now store the remote URI for an event in the COMMENTS for the
event.  mKCal uses a heuristic to split comments which unfortunately
fails for calendar path URIs which contain a space.

This commit ensures that when the remote calendar path contains a
space (or a percent) we percent-encode the path before storing it
in the comment, and that we decode the path before returning it.

Contributes to MER#982