[buteo-sync-plugin-caldav] Deletions of series should invalidate exception deletions. Contributes to MER#1195
When an entire series is deleted, any persistent exceptions for that
series are also marked as deleted.  However, attempting to upsync
those deletions will fail once the original series is deleted, since
the URI will no longer exist.

This commit ensures that we don't attempt to upsync deletions of
persistent exceptions to already-deleted series.

Contributes to MER#1195